Use common method for action explorer and api form menu
