+++ app/client/src/pages/Editor/Explorer/Actions/ActionEntityContextMenu.tsx
@@ -1,9 +1,8 @@
-import React, { useCallback, useMemo } from "react";
+import React, { useCallback } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import TreeDropdown from "components/editorComponents/actioncreator/TreeDropdown";
 
 import { AppState } from "reducers";
-import { getNextEntityName } from "utils/AppsmithUtils";
 import ContextMenuTrigger from "../ContextMenuTrigger";
 
 import {
@@ -14,30 +13,8 @@ import {
 
 import { initExplorerEntityNameEdit } from "actions/explorerActions";
 import { ContextMenuPopoverModifiers } from "../helpers";
-import { groupBy, noop } from "lodash";
-import { ActionData } from "reducers/entityReducers/actionsReducer";
-
-const useNewAPIName = () => {
-  // This takes into consideration only the current page widgets
-  // If we're moving to a different page, there could be a widget
-  // with the same name as the generated API name
-  // TODO: Figure out how to handle this scenario
-  const actions = useSelector((state: AppState) => state.entities.actions);
-  const groupedActions = useMemo(() => {
-    return groupBy(actions, "config.pageId");
-  }, [actions]);
-  return (name: string, destinationPageId: string) => {
-    const pageActions = groupedActions[destinationPageId];
-    // Get action names of the destination page only
-    const actionNames = pageActions
-      ? pageActions.map((action: ActionData) => action.config.name)
-      : [];
-
-    return actionNames.indexOf(name) > -1
-      ? getNextEntityName(name, actionNames)
-      : name;
-  };
-};
+import { noop } from "lodash";
+import { useNewActionName } from "./helpers";
 
 type EntityContextMenuProps = {
   id: string;
@@ -46,7 +23,7 @@ type EntityContextMenuProps = {
   pageId: string;
 };
 export const ActionEntityContextMenu = (props: EntityContextMenuProps) => {
-  const nextEntityName = useNewAPIName();
+  const nextEntityName = useNewActionName();
 
   const dispatch = useDispatch();
   const copyActionToPage = useCallback(
