+++ app/client/src/pages/Editor/Explorer/Actions/MoreActionsMenu.tsx
@@ -2,7 +2,6 @@ import React, { useCallback } from "react";
 import { useDispatch, useSelector } from "react-redux";
 
 import { AppState } from "reducers";
-import { getNextEntityName } from "utils/AppsmithUtils";
 
 import {
   moveActionRequest,
@@ -13,18 +12,7 @@ import {
 import { ContextMenuPopoverModifiers } from "../helpers";
 import { noop } from "lodash";
 import TreeDropdown from "components/ads/TreeDropdown";
-
-const useNewAPIName = () => {
-  // This takes into consideration only the current page widgets
-  // If we're moving to a different page, there could be a widget
-  // with the same name as the generated API name
-  // TODO: Figure out how to handle this scenario
-  const apiNames = useSelector((state: AppState) =>
-    state.entities.actions.map((action) => action.config.name),
-  );
-  return (name: string) =>
-    apiNames.indexOf(name) > -1 ? getNextEntityName(name, apiNames) : name;
-};
+import { useNewActionName } from "./helpers";
 
 type EntityContextMenuProps = {
   id: string;
@@ -33,7 +21,7 @@ type EntityContextMenuProps = {
   pageId: string;
 };
 export const MoreActionsMenu = (props: EntityContextMenuProps) => {
-  const nextEntityName = useNewAPIName();
+  const nextEntityName = useNewActionName();
 
   const dispatch = useDispatch();
   const copyActionToPage = useCallback(
@@ -42,7 +30,7 @@ export const MoreActionsMenu = (props: EntityContextMenuProps) => {
         copyActionRequest({
           id: actionId,
           destinationPageId: pageId,
-          name: nextEntityName(`${actionName}Copy`),
+          name: nextEntityName(`${actionName}Copy`, pageId),
         }),
       ),
     [dispatch, nextEntityName],
@@ -54,7 +42,7 @@ export const MoreActionsMenu = (props: EntityContextMenuProps) => {
           id: actionId,
           destinationPageId,
           originalPageId: props.pageId,
-          name: nextEntityName(actionName),
+          name: nextEntityName(actionName, destinationPageId),
         }),
       ),
     [dispatch, nextEntityName, props.pageId],
