+++ app/client/src/pages/Editor/Explorer/Actions/helpers.tsx
@@ -1,4 +1,4 @@
-import React, { ReactNode } from "react";
+import React, { ReactNode, useMemo } from "react";
 import { apiIcon, dbQueryIcon, MethodTag, QueryIcon } from "../ExplorerIcons";
 import { PluginType } from "entities/Action";
 import { generateReactKey } from "utils/generators";
@@ -15,6 +15,11 @@ import { ExplorerURLParams } from "../helpers";
 import { Datasource } from "entities/Datasource";
 import { Plugin } from "api/PluginApi";
 import PluginGroup from "../PluginGroup/PluginGroup";
+import { useSelector } from "react-redux";
+import { AppState } from "reducers";
+import { groupBy } from "lodash";
+import { ActionData } from "reducers/entityReducers/actionsReducer";
+import { getNextEntityName } from "utils/AppsmithUtils";
 
 export type ActionGroupConfig = {
   groupName: string;
@@ -134,4 +139,26 @@ export const getPluginGroups = (
       />
     );
   });
+};
+
+export const useNewActionName = () => {
+  // This takes into consideration only the current page widgets
+  // If we're moving to a different page, there could be a widget
+  // with the same name as the generated API name
+  // TODO: Figure out how to handle this scenario
+  const actions = useSelector((state: AppState) => state.entities.actions);
+  const groupedActions = useMemo(() => {
+    return groupBy(actions, "config.pageId");
+  }, [actions]);
+  return (name: string, destinationPageId: string) => {
+    const pageActions = groupedActions[destinationPageId];
+    // Get action names of the destination page only
+    const actionNames = pageActions
+      ? pageActions.map((action: ActionData) => action.config.name)
+      : [];
+
+    return actionNames.indexOf(name) > -1
+      ? getNextEntityName(name, actionNames)
+      : name;
+  };
 };
